165 | Rebase onto dev branch
Rebased onto dev which included...

* dnatracing
* combining grainstats with dnatracing stats
* fail gracefully
* May or may not handle single grains correctly, need to reconcile with new threshold methods.

I've removed all plot_and_save() from Filters/Grains classes and that is all done at the end. Filters plots go into <output>/<filename> and because Grains can be on upper and lower thresholds their plots go into a further subdirectory for each threshold (i.e. <output>/<filename>/[upper|lower]).

Many testes now fail (13 failed, 30 passed, 26 errors) but this is an artefact of not having yet reworked the tests to handle the new threshold methods that are used in filtering and the big change in the structure of the Classes. It may not be as bad as it first seems to repair these.