241 | Fixing missing lower statistics
Couple of problems related to not concatenating the dnatracing statistics for lower and upper.

Then merging grainstatistics with dnatracing statistics was only done on molecule rather than accounting for the
threshold.

Somehow the resetting of index within dnatracing was no longer done in place either.

Now get 22 molecules (21 upper, 1 lower) as before.