Regression test for run_topostats entry point
Closes #308

Introduces a couple of extra tests...

* One that checks the help is displayed when `run_topostats [-h|--help]` is run.
* One that checks `run_topostats` runs completely using the `topostats/default_config.yaml`.

I couldn't work out how to get to run using the package resources because if no arguments are added then because the
call to `run_topostats` is made within `pytest` the `sys.argv` that `pytest` was invoked with are instead passed and so
this errors. Instead we explicitly specify the configuration file with the `--config topostats/default_config.yaml`.

For a better explanation of this situation see [this thread](https://stackoverflow.com/a/55260580/1444043).