Improve logging verbosity configuration
Started investigating #381 and wanted to use LOGGER.debug() but hadn't configured this correctly in the first
instance. This sets up the `LOGGER.setLevel` conditional on the configuration value of `log_level` and introduces the
command line option `-l/--log_level`. At the same time it negates the need for `warnings` and `quiet` in the
configuration.

Validation and documentation updated to reflect this change.

When working on code and debugging you can now use `LOGGER.debug(f"[{self.filename}] [<function/method>] <MESSAGE>")`
instead of `print(<MESSAGE>)` and see the output by using `run_topostats -l debug`.