_count_neighbours and _inverse_mask; split conftest
Counting neighbours that are non-zero, deriving inverse masks for skeletons so we only count the number of ends (cells
with only 1 non-zero cell adjacent) on the skeleton. Lots of tests for each of these.

This method of counting neighbours does actually have a problem so isn't reliable (see diagonal1 expected)