Initial determination of circular shape
Currently uses one of two methods `ends` or `regions` to determine whether molecules are circular or not.

This works in a sense but there are commonly covalently bound molecules which appear to be loops but are not
closed molecules in strict sense.

Part of the problem will involve pruning as small branches can appear on true circular molecules and so counting "ends"
is problematic.