227 | Refactoring plotting (and more)
Closes #227 #451 #452

Refactoring plotting.py

This is the initial refactoring of `plotting.py`. Because of the vast scope for variation that exists in generating
plots a deliberate attempt to keep a sensible and minimal set of plots has been taken whilst refactoring the code to be
as succinct as possible. Originally there was a mixture of Matplotlib and Seaborn functions, to maintain simplicity the
refactoring uses only Seaborn.

* Histograms and Kernel Density plots can be generated (if both are requested then the KDE is overlaid).
* Violin plots can be generated.
* For either type of plot if there is more than one unique `image` in the dataset the data is plotted separately on the
same graph (i.e. different colour histograms/kde for each image; multiple violins, one for each image).
* If data comes from more than one image then these are plotted separately by default (currently there is no option to
combine them) and Seaborn's `hue` option is used to differentiate these. Because this `hue` option is always used, even
with a single image, it overrides the `color` option so it is not possible to explicitly request the color of plots.
* Configuration is via a new YAML file `topostats/summary_config.yaml`. When running as part of `run_topostats` output
is to the `default_config.yaml` `output:` location but in the sub-directory `summary_distributions/`.
  * Output file type can be controlled via the `file_ext:` argument.
  * Output file names are constructed as `<statistic>_<plot_type>`.
  * Statistics to plot are controlled by (un-)commenting the variables listed under `stats_to_sum` in the
  `sumary_config.yaml`.
* Variable names (which are currently aiming to all be `lower_snake_case`) are automatically mapped to variable labels
via `topostats/var_to_label.yaml`. New statistics/variables when introduced should be added to this file to ensure plots
are labelled correctly.
* Plotting will be done automatically when running TopoStats unless `default_config.yaml` has the option `summary_stats:
run: false`. Unless an explicit configuration file for summary_stats (i.e. `summary_config.yaml`) is specified in
this section of `default_config.yaml` then the default `topostats/summary_config.yaml` is used.
* An "entry-point" is provided to run plotting retrospectively on generated CSV files. It also includes the option
`--create_config_file` that, in a similar vein to `run_topostats --create_config_file` will write a sample YAML
configuration file to the specified location for subsequent editing. It can then be passed to `toposum --config_file
custom_config.yaml` allowing the user to have an easy way to toggle configuration options.
* An option exists to save the all generated images as [Python
pickles](https://realpython.com/python-pickle-module/). This can then be loaded in a Jupyter Notebook and the images
further modified.
* Documentation has been updated under `configuration.md` and `usage.md`.
* A sample Jupyter Notebook has been introduced (`notebooks/01-Summary-statistics-and-plots.ipynb`) that shows how to
load a CSV file of statistics and use Pandas and Seaborn to create these images. I think it would be really useful to
see this grow with examples of possible plots that may be generated. The file is version controlled and can therefore be
added, but its important to restart the kernel and clear cells before committing otherwise output (text and graphs) are
included.

ToDo (maybe?)

Rather than this work growing any larger I've deliberately left some (small?) tasks undone, maybe not all of them need
doing though. Once this PR is merged these can be discussed and written up as individual issues and addressed.

* **A per image summary of statistics**. This is pretty straight-forward, the `self.df` is a Pandas Dataframe to
summarise it you can `self.df.groupby("image").describe()` and a summary dataframe is returned. Might not include all
metrics (e.g. median, inter-quartile range) but they can also be generated in a similar manner and merged.
* **Automatically Adding Units** - Most units are `nm`, but area and volume differ as to statistics of
curvature. Probably the easiest approach will be to modify the labels in `var_to_label.yaml` so that titles carry the
correct unit. However it may be desirable to have axes just show the unit and the title not by default. If this is the
case the expanding `var_to_label.yaml` so that each `var` has a nested dictionary of `title` and `units` will be the
most sensible and cleanest approach.
* **Multiple CSV files** - the entry-point "toposum" should have additional functionality added to scan a `base_dir` for
CSV files and these should be read in and combined into one Pandas DataFrame and passed on to the `toposum()`
function. This will require adding a `base_dir: ./` option to the configuration file (`summary_config.yaml`) and
associated validation to `topostats/validation.py`. The existing `topostats.io.find_files()` function (nee
`find_images()` see below) should be used to find CSV files. Once a list of files exists its possible to use dictionary
comprehension in conjunction with `pd.concat()` to read and append all files in one go (e.g. `pd.concat([pd.read_csv(x)
for x in csv_files])`).
* **Plotting all molecules across grains** This is another variation that would require a more logic and duplication of
code. I think if this is desired it would be best suited to generating via Jupyter Notebooks.
* **Examples of modifying pickled images** It might have some utility to show how to load the pickled images

Other changes included

House Keeping

* Introduces `flake8-print` to the `dev` requirements so that extraneous `print()` statements from debugging are
detected by `pre-commit` and removed before Pull Requests are made (#452).
* Removes `@pytest.mark.mpl_image_compare()` from `test-*-minicircle.py` tests. The various steps in processing assess
the Numpy arrays for consistency by taking their `sum()` whilst I initially incorporated `pytest-mpl` to compare the
plots these would generate that was perhaps over-zealous and unnecessary. It also results in a lot of minor changes in
borders requiring the images to be regenerated (see #451). They have therefore been removed along with the associated
`tests/resources/img/*.png` files. Tests for `test_plottingfuncs.py` remain in place with Matplotlib graphs generated.
* Adds `.git-blame-ignore-revs` should we need to ignore blame (see [Who's to
blame](https://rse.shef.ac.uk/blog/git-blame/) and includes a commit made automatically by `pre-commit.ci` which applied
`black` to existing files. We don't want the blame to be attributed to a CI script.

Moving/renaming functions

* `topostats/run_topostats.py` was getting big and unwieldy. The `process_scan()` function has therefore been moved to
its own sub-module `topostats/processing.py` and is imported to `run_topostats.py` for usage. Tests have been moved as
is appropriate.
* Renamed `find_images()` > `find_files()` since it was requested that the new `toposum` command be able to load
multiple CSV files (a feature not yet implemented).
* Command line options to `run_topostats` that are highly unlikely to be used have been removed.