Correct --create-label-file and add test
Thanks to @SylviaWhittle for thorough testing this wasn't correct, the whole config file was written rather than the
`var_to_label` dictionary. That is now corrected and a test introduced to check this is written correctly.

This required some additions to `plotting.py` to work within `pytest` including adding an argument option to specify the
CSV file that is to be summarised.