Makes DPI configurable via default_config.yaml
Closes #526

The `Images` class always had `dpi` as an option but it was not accessible from `run_topostats` as no options were
present in the configuration file and therefore didn't cascade through. Dots Per Inch (DPI) can now be specified in the
configuration file, its default is set to `100.0` (the default used by `savefig` in Matplotlib). Validation is performed
and a test introduced that sets `dpi = 300` for testing to check the value cascades through and plots are made using
this value (this also required setting the parameter for the `pytest-mpl` fixture too, otherwise it would save the
figure with the default value).