Improve updating of plot_dict
In adding `dpi` as an option I forgot to update the individual plotting dictionaries with this new option. To avoid such
errors in the future I've changed how these dictionaries are updated to expand all options (bar `plot_dict` which would
be recursive and `run` which is not required when plotting individual images).