Add sphinx-autoapi; removing pre-generated API .rst files
Closes #532

The problem was that [sphinx-multiversion](https://holzhaus.github.io/sphinx-multiversion/master/configuration.html)
checks out each branch/tag that it is building for prior to build documentation. If there are no `.rst` files in that
branch/commit for API documentation then the API docs can not be generated.

The solution is to use [sphinx-autoapi](https://sphinx-autoapi.readthedocs.io/en/latest/) to automatically generate the
`.rst` files for the API on each run that `sphinx-multiversion` makes on each branch/tag it is building documentation
for.

Also sorts package dependencies in `pyproject.toml`.