Empty dataframe when no grains found in image
Closes #514

Further to #515 this PR changes the structure of the dataframes that are returned by `create_empty_dataframe()` which
originally had a row `NaN` when no grains were found in a given image. This led to the situation where it looks like an
image has successfully processed simply because the `image_name` was in the final `results` DataFrame.

Instead a genuinely empty data frame is now generated with just column headers on a per-image basis that can be appended
to results from images that have had grains successfully detected.

Other tweaks...

* In addition I've added an additional line to the `config.yaml` that is generated pointing people to the documentation
on configuration files/parameters.