Corrects alignment/merging of grainstats and tracing_stats dataframes
Closes #403

If DNA tracing removed grains because the skeleton was less than a hard coded threshold (10) in size then no tracing
statistics were calculated for that grain (as it was dropped).

This threshold is now configurable (via `crap_threshold` since its the `purge_obvious_crap()` method that performs
this).

The way molecules are currently numbered always starts at `1` because the _whole_ image with grains represented as masks
is taken as input. `0` is the background and therefore removed from the molecule numbers which therefore started at
`1`. In contrast `grainstats_df` indexed molecules from `0`. This resulted in mis-alignment when merging.

Further when merging `tracing_stats_df` and `grainstats_df` the merge was performed based on `how="inner"` and so only
those molecule numbers common to both were retained. This has been corrected and `how="left"` is used in preference
where the "left-most" dataframe is the `grainstats_df`. This means that even if a grains is dropped from tracing its
grain statistics are retained in output. A logging message is reported to indicate how many grains are removed.  Tests
are included which result in two grains being removed and checks are made to ensure that there are not tracing
statistics for these grains.

The `basename` is no longer added to the `trace_stats_df` and is now done after merging with `grainstats_df`.

I would be very grateful for a careful review of this please @JeanDu as reporter of the issue so that it can be included
in the pending release, toggling `crap_threshold` is one way to force some grains to be dropped from DNA
tracing. Naturally other reviews are of course welcome.