Initial dnatracing refactor
Work to date on refactoring DNAtracing. This introduces `topostats.tracing.trace_dna` sub-module with a class
`traceDNA`.

This is far from complete! None of the code is currently used _anywhere_ in processing, but that is ok, the code can sit
there and wait until it is complete with comprehensive tests and then we can switch out the
`topostats.tracing.dnatracing` sub-module and its classes with these new ones.

I want to avoid asking people to review lots of code (and there is already a lot in here!) and its important that I get
feedback regularly.

* Currently in the classes within `topostats.tracing.dnatracing` the _whole_ image is processed. `traceDNA` class takes
a different approach and the unit of processing is now an individual grain. This avoids having to repeatedly loop over
all found grains and in turn avoids the issues that @Jean-Du encountered with statistics not aligning. When it comes
time to process a whole image each grain is passed in and the results returned (or `None` if it fails to meet
thresholds/elsewhere).
* The `dnaTrace` class takes the approach of finding the co-ordinates of each skeletonised molecule and working out if
it has an adjacent point to count how many ends there are. I've used a method which leverages NumPy's vectorisation and
have tried to explain this in the docstring.
* I'm trying to take a Test Driven Development approach and write tests before the code (or if failing in conjunction
with). These are heavily parameterised because there are lots of possibilities for molecules but they use small dummy
examples rather than large observed skeletons. Some of these may not (yet) have been observed in real data, but that
doesn't preclude them from cropping up in the future. The aim is to test ~all~ as many eventualities as I can think of
up-front during development.
* Switched the `get_skeleton()` function to a class.
* `topostats.tracing.utils` sub-module is very much a  work in progress, it holds skeleton classes and methods at the
moment.