[BUG] get_fitted_traces was not extracting heights correctly
The perp_array is the range around a skeleton point where the heights are obtained from the original image.
The co-ordinates for skeletons are obtained using np.argwhere() which returns a two-dimensional array of the `[x, y]`
co-ordinates. These co-ordinates are used on line 331 of tracing/dnatracing.py   to get the heights and that currently
in the main branch reads...

```
height_values = self.gauss_image[perp_array[:, 1], perp_array[:, 0]]
```

But when indexing arrays its the row that comes first not col and so the above line tries to pull out the heights based
on the `y, x` co-ordinates rather than the `x, y` co-ordinates which is why we observed IndexError. This probably didn't
happen very often at all with the original dnaTrace  class because the whole image is passed in and so typically this
problem would not be encountered. Correcting the above line and including a try: ...except: to catch errors where we are
near the boundary and if so set that to be the boundary....

```
try:
    height_values = self.gauss_image[perp_array[:, 0], perp_array[:, 1]]
except IndexError:
    perp_array[:, 0] = np.where(
        perp_array[:, 0] > self.gauss_image.shape[0], self.gauss_image.shape[0], perp_array[:, 0]
    )
    perp_array[:, 1] = np.where(
        perp_array[:, 1] > self.gauss_image.shape[1], self.gauss_image.shape[1], perp_array[:, 1]
    )
height_values = self.gauss_image[perp_array[:, 1], perp_array[:, 0]]
```

..has I think solved the problem.

This commit also introduces padding of the cropped image and mask which are likely to be required in subsequent newer
processes. This is instead of padding the bounding box. If images are near the border of the image and padding goes
beyond the boundaries then the extent of the image is retained (need to write tests for this!).