Improves processing of multiple grains during dnatracing
This removes processing of a whole image and instead makes the unit of processing in dnatracing a single
grain. Processing multiple grains is done via a helper function `trace_image()` which calls `trace_grain()` the later is
a thin wrapper to the `traceDna()` class.

When cropping the image and masks for processing they are padded twice compared to the scikit-image determined bounding
box.

1. The original image/mask are paded by a user configurable amount. This is so that heights which will be used when
height-based tracing is introduced are available outside of the region.
2. These images are then padded further to allow some grains that are close to the edge/border of images to be processed.

Currently the value for both steps of padding is the same and is set to `20`, other values may be more appropriate.

Two bugs were identified during this work and have (hopefully) been corrected. See commits 236750b2 and 2a79c4ff for
further details.

Summary of changes.

* Updates tests for test_tracing_methods to work with dnatracing now loops have been removed.
* Uses loops to pass over all grains in tracing because can't run parallel process from within a parallelised
  process.
* Code left in place for running in parallel when we get round to modularising running TopoStats properly.
* Adds a new function for preparing arrays (`prep_arrays()`) along with tests.
* Removes the class TraceStats completely no longer required
* Removes medial as a method of skeletonisation
* Some (but not all) docstrings added (most will get added during subsequent refactoring I hope).
* Updates target regression test files.