Correct Loading gwy channels and units
# XY, Z units correction
* units for px_to_nm to multiply data array should be picked up from 'si_unit_z' instetad of 'si_unit_xy'.
* Non-square image should be considered. Probably, px_to_nm_z, px_to_nm_x, px_to_nm_y will be necessary.

# Channel index
Data key of gwy sometimes starts from 1 or larger. (ex. '/4/data') It happens when image is cropped on Gwyddion, and it remains after saving even other channels are removed.

A loop to scan indices was added. And it guesses whether if the channel is topography or not by si_unit_z.