Moves DPI configuration from default_config to plotting_dictionary
Closes #645

Moves specification of a global DPI for all images from `topostats/default_config.yaml` to per-plot in
`topostats/plotting_config.yaml` with default of `100` for most values but `800` for existing plots of traces.

Updates validation rules and documentation and I've taken the opportunity to add missing Description to two rows in the
documentation covering the `summary_config.yaml`.