Add snoop as a dependency but disable by default
I came across [snoop](https://github.com/alexmojaki/snoop) which provides a powerful set of debuggin tools and think it
would be useful to use it in TopoStats.

This PR adds it as a `dev` optional dependency and imports and installs during package initialisation. By default
thought it is disabled so that performance in not impacted, this is done by `snoop.install(enabled=False)` in
`__init__.py`.

To enable and using during development add the `@snoop` decorator to a function, method or class and in the same file
re-enable dynamically by calling `snoop.install(enabled=True)`.