Add environment to pypi.yaml publishing workflow
PyPI have introduced the option to publish from trusted environments using [OpenID
Connect](https://docs.pypi.org/trusted-publishers/adding-a-publisher/).

Following the [instructions](https://docs.pypi.org/trusted-publishers/adding-a-publisher/) it is recommended to include
an `environment` under which the workflow runs.

To this end `pypi-publish` environment has been created for the TopoStats repository and this PR adds the `environment:
pypi-publish` configuration to the `jobs.build-release`. At the same time a trusted publisher has been
[added](https://docs.pypi.org/trusted-publishers/adding-a-publisher/) to the `topostats` account on PyPI.