PR Feedback
Thanks for the feedback @MaxGamil-Sheffield this commit...

+ Loosens validation of `cmap` so that any Matplotlib color map can be used.
+ Removes reporting of DPI/output format/cmap from early logging stages and output of `completion_message()`.

I hadn't thought about `None` being listed in the `completion_message()` for DPI/Output Format/cmap and appreciate this
would be confusing so thanks for highlighting that. The solution I've gone for (removing the additions that reported
these) is different from that suggested (update the `config` dictionary with parameters from `mpl.rcParams` early in
processing). My reasoning being...

+ Previously we didn't report these, no one has ever asked to see them in the logging output.
+ We write configuration options to YAML file via `write_yaml()` at the end of processing. Its a verbatim copy of that
  which was used (either user specified or `default_config.yaml`)and it contains the settings used. If a user didn't
  specify DPI/cmap/format I'm not sure we should alter this. It could be argued it is useful to provide them but then
  that would also require writing _all_ other configuration/plotting options to be consistent should the default values
  ever change in the future.

Currently only a handful of parameters are read from `topostats.mplstyle` and this is conditional on
whether any of these are being over-ridden or not when instantiating the `Images()` class. Currently we

```
plottingfuncs.Images() > load_mplstyle() > Set DPI/cmap/format based on arguments to Images()
```

It is certainly possible to change this process as suggested and..

```
load_mplstyle() > Update DPI/cmap/format > plottingfuncs.Images()
```

...but that is a larger amount of work to undertake and introduces scope drift to this PR. If it is desirable to report
this information in logging and/or ensure the configuration file that is written contains the default parameters from
`topostats.mplstyle` then we can address that as a separate issue.