Switch to vector projection method for triangle heights
The inverse gradient results in `ZeroDivisionError` when the base of the triangle is horizontal so switching
methods (thanks for the comprehensive suggestions of both @SylviaWhittle).

In order to check everything I wrote a plotting function (see previous commit) and whilst many of the minimum feret
distances are correct there are still instances where the triangle height is calculated _outside_ of the convex hull
which is baffling. For the most part not an issue but in the `holo_ellipse_angled` this **is** an issue.

Figures below show the problem.

The same Graham Scan algorithm is used for the rotating caliper pairs as in GrainStats so unsure if this is an issue
there and can't check as co-ordinates aren't currently returned (possibly bounding box may be informative as I think I
saw that was calculated for the feret distances, but inclined to solve this in current code base).