Exclude triangle heights (min feret) if outside hull
The minimum feret which is calculated as the height of the triangle formed between the rotating caliper pair and the
next adjacent point on the hull (with the apex being the antipodal point on the opposite hull) resulted in heights with
co-ordinates that were outside of the hull if the triangles were scalene.

In turn these gave incorrect minimum feret coordinates, even if the distances were correct.

To avoid this a check is now made as to whether the line formed by the triangle height is `in_polygon()` using functions
from Shapely.