Ensures target directories exist
In #806 and refactoring (see 07db71820) I found an unintended side-effect was that some of the target
directories (specifically `filter_out_path` and `grain_out_path`) were no longer present and the
tests (`tests/test_processing.py`) failed.

These are meant to be explicitly created by `processing.get_out_paths()` but for some reason were now no longer.

For filters the solution seemed counter-intuitive as the images are saved to `filter_out_path` if the image is part of
the `core_set`, otherwise they are saved to `core_out_path`.

For grains the `direction` is appended to the `grain_out_path` once (rather than three times) and the `.mkdir(parents =
True, exists_ok=True)` pathlib method is used to ensure the directory exists prior to attempting to plot.