tests(pruning): Tests TopoStatsPrune and convPrune
+ Adds a smaller (30x30) test skeleton for testing pruning, this makes it easier to work with, however, this does raise
  a question as the resulting skeleton, generated by `skimage.morphology.skeletonize()` does have a T-shaped junction
  which it has been suggested never arise (more on this below).
+ Tests for `TopoStatsPrune` and `convPrune` attempt to demonstrate the effect of varying parameters has on pruning to
  which end the following parameters are varied...

  + `max_length`
  + `height_threshold`
  + `method_values`
  + `method_outliers`

There are some issues identified so far...

`convPrune`

As noted in the `reason` for the first paramterised test failing which arises when `max_height = None` (since `int` can
not be compared to `None`) I think there is a problem with the default height being used here.

Line 454 tries to set the default max_branch_length to be 15% of the total number of elements in the skeleton
array. `self.skeleton.size` and counts the number of elements in an array (and because we are working with 2D images
this will always be the product of the number of rows and columns). But the length of this will _always_ be `1` since
the total number of elements in the skeleton will always be an integer. Thus if `max_length == -1` then the
`max_branch_length` will always be 0!

That said in this dummy example we don't even get that, instead we get an error raised

```
max_length = -1
simple_array = np.asarray([[0, 0], [0, 0], [0, 0]])
total_points = simple_array.size
max_branch_length = max_length if max_length != -1 else int(len(total_points) * 0.15)\

TypeError: object of type 'int' has no len()
```

Its also unclear why the `convPrune._prune_by_length()` mtehod takes the argument `max_length` when its an attribute of
the class which could be used instead.

This doesn't happen in the `topostatsPrune` class because the length of the molecule is based on the co-ordinates of the
skeleton rather than the size of the array.

Excessive Complexity/Code duplication

The splitting of functionality here into multiple classes means there is some code duplication. Both `topostatsPrune`
and `convPrune` have methods to `_prune_by_length()` conditionally based on the value of `self.max_length` not being
`None` but they do so in different manners. Both methods then go on to called `heightPruning()` in the same manner.

The two classes could be combined into one with an option of whether to use the original length based pruning or the
convolutional approach and call the appropriate method (renaming the method from `convPrune`  to
`prune_length_by_conv()` and the one from `topostatsPrune` to `prune_length_by_neighbours()`. In turn this would mean
that there is no need to have the `pruneSkeleton()` factory class (and having that as a class in itself seems like
overkill when a series of functions would suffice and address the complaints from Pylint on `too-few-public-methods`
too).

Handling multiple grains

The refactoring done previously in #600 removed the loops from every method/function in the tracing module so that it
only handled a single grain. The looping over of multiple grains is handled by code in `process_scan.py`, this means the
`prune_all_skeletons()` methods can be removed too, further simplifying the code base.

T-shaped junctions

Previous work in #835 originally tested whether `pruning.rm_nibs()` "Remove single pixel branches (nibs) not identified
by nearest neighbour algorithsm as there may be >2 neighbours" and the current tests show that these are still left by
the pruning methods being tested, even though the last step is to re-run the plain skeletonisation method from
`skimage` (Zhang's).

The following is a reproducible example that shows such a nib remains after pruning

```python

import numpy as np
import matplotlib.pyplot as plt
from skimage import draw, filters
from skimage.morphology import skeletonize

def _generate_heights(skeleton: npt.NDArray, scale: float = 100, sigma: float = 5.0, cval: float = 20.0) -> npt.NDArray:
    """Generate heights from skeletons by scaling image and applying Gaussian blurring.

    Uses scikit-image 'skimage.filters.gaussian()' to generate heights from skeletons.

    Parameters
    ----------
    skeleton : npt.NDArray
        Binary array of skeleton.
    scale : float
        Factor to scale heights by. Boolean arrays are 0/1 and so the factor will be the height of the skeleton ridge.
    sigma : float
        Standard deviation for Gaussian kernel passed to `skimage.filters.gaussian()'.
    cval : float
        Value to fill past edges of input, passed to `skimage.filters.gaussian()'.

    Returns
    -------
    npt.NDArray
        Array with heights of image based on skeleton which will be the backbone and target.
    """
    return filters.gaussian(skeleton * scale, sigma=sigma, cval=cval)

def _generate_random_skeleton(**extra_kwargs):
    """Generate random skeletons and heights using skimage.draw's random_shapes()."""
    kwargs = {
        "image_shape": (128, 128),
        "max_shapes": 20,
        "channel_axis": None,
        "shape": None,
        "allow_overlap": True,
    }
    # kwargs.update
    heights = {"scale": 100, "sigma": 5.0, "cval": 20.0}
    kwargs = {**kwargs, **extra_kwargs}
    random_image, _ = draw.random_shapes(**kwargs)
    mask = random_image != 255
    skeleton = skeletonize(mask)
    return {"original": mask, "img": _generate_heights(skeleton, **heights), "skeleton": skeleton}

def pruned_plot(gen_shape: dict) -> None:
    """Plot the original skeleton, its derived height and the pruned skeleton."""
    img_skeleton = gen_shape
    pruned = topostatsPrune(
        img_skeleton["img"],
        img_skeleton["skeleton"],
        max_length=-1,
        height_threshold=90,
        method_values="min",
        method_outlier="abs",
    )
    pruned_skeleton = pruned._prune_by_length(pruned.skeleton, pruned.max_length)
    fig, ((ax1, ax2), (ax3, ax4)) = plt.subplots(2, 2)
    ax1.imshow(img_skeleton["original"])
    ax1.set_title("Original mask")
    ax2.imshow(img_skeleton["skeleton"])
    ax2.set_title("Skeleton")
    ax3.imshow(img_skeleton["img"])
    ax3.set_title("Gaussian Blurring")
    ax4.imshow(pruned_skeleton)
    ax4.set_title("Pruned Skeleton")
    plt.show()

def pruning_skeleton() -> dict:
    """Smaller skeleton for testing parameters of prune_all_skeletons()."""
    return _generate_random_skeleton(rng=69432138, min_size=15, image_shape=(30, 30))

pruned_skeleton(pruning_skeleton())
```

The paramterised test with `id="Length pruning enabled (10), removes two ends leaving nibs on both."` includes the
resulting array after pruning and there is a two-prunged fork of nibs when using `convPrune()`.

I think it would be useful if `remove_nibs()` were able to handle such T-junctions as well.