test(tracing): Skip tests in test_dnatracing_single_grain.py
Work on getting the test suite to work with the refactored code. Long way off working at the moment though.

+ Pass `default_config.yaml` into the `dnatrace_linear`/`dnatrace_circular` fixtures so we have all options in place.
+ Skip a bunch of tests that are contingent on the pruning having been done correctly, this will require tweaking of the
  `skeletonise_params` so that correctly pruned skeletons are returned.
+ `self.grain` > `self.mask` I found it confusing that the original image after smoothing was called
  `self.smoothed_grain` when it was heights and the mask was called `self.grain`. By having the mask named as such it
  avoids confusion and makes the attributes clearer.

A major hindrance to debugging this is when the skeleton is converted from an array to a list of co-ordinates that
define the location of points.

This makes it a bit of a pain to then plot and see what we have can be achieved with the following messy code...

```
original_skeleton_coords = np.asarray([[28, 47],
       [ 28,  48],
       [ 29,  46],
       [ 29,  48],
       [ 30,  47],
])
skeleton = np.zeros((original_skeleton_coords.max() + 2, original_skeleton_coords.max() + 2))
skeleton[original_skeleton_coords[:,0], original_skeleton_coords[:,1]] = 1
plt.imshow(skeleton)
plt.show()
```