refactor(dnatarcing) : Remove obsolete gaussian_filter()
The `dnaTrace.gaussian_filter()` method is obsolete and has been replaced by `dnaTrace.smooth_mask()`. The obsolete
function has therefore been removed and the related test adapted and expanded to test `test_smooth_mask()` with
parameters that vary the number of `dilation_iterations` and the value for `gaussian_sigma`.

The new `dnaTrace.smooth_mask()` performs binary dilation of the mask and compares this to a gaussian blur of the
mask. Whichever results in the smallest difference compared to the original mask then has "holes readded" and the
`self.smoothed_mask` is updated.

It is important to note that this differs from previous methods where the Gaussian filter was applied to the original
images of heights. Why this switch was made is unclear.

Additional work

+ Refactoring to remove mention of `grains` as the term is used in multiple places, sometimes with the original height
  array and sometimes with the mask array. This ambiguity can lead to confusion instead the terms `image` and `mask` are
  used.