bug(height_profiles): correctly interpolate points
One of the tests was deliberately skipped and looked weird as only three points were interpolated. On starting to
incorporate into the main workflow tests revealed why this is. The feret coordinates were being sorted independently
which was a mistake, it decoupled and often flipped over the points between which interpolation was occurring.

Further we were always taking the x-axis and interpolating along that but when the y-axis was the longer distance this
resulted in sparse interpolation. Instead we now perform linear interpolation on the axis with longest difference first
and then interpolate the points on the other axis to determine the points at which we want to interpolate heights.