wip: Adding all configuration options as arguments
- Renames `toposats.run_topostats` > `topostats.run_modules` as its more descriptive of what the submodule is doing.
- Moves generic command line options to be arguments of the parser.
- Adds arguments to each sub-parser for all module options.
- Adds tests to check that each argument is correctly returned.
- Updated and made a few names consistent.

This is still a Work In Progress but it shouldn't break anything (tests have all passed locally).

Whether to merge this depends perhaps on the desire to tag and make a release and whether there is anything else pending
that is considered essential to be in that release (e.g. documentation). If so then _don't_ merge this.