refactor(unet): Square bounding box based on center and radius
Alternative approach to squaring a bounding box.

First and foremost, only attempt to do this if the existing bounding box isn't square.

That said its not a great solution because not all tests pass because using `np.floor()` and `np.ceil()` on `center -/+
radius` doesn't always give the desired square box.

It may be worthwhile cherry-picking some specific files from this commit though as I've updated the `id=""` in the test
and added the clause to `grains.py` which only calls `square_bounding_box` (nee `make_bounding_box_square`) if the
bounding box isn't square in the first place which might reduce the number of calls made.