style(pylint): Apply some linting from pylint
Applied some of the linting rules that `pylint` reported but not all, the environment used doesn't have TopoStats
installed and so fails (just as `pylint` would fail in `pre-commit.ci` and is therefore configured as `local` only).

Have opted to remove `nbqa-pylint` hook but have removed unused imports from Notebooks.