feature: Adds --image-set as option to topostats
Closes #1057

In reviewing #1056 I wanted to look at the individual grain curvature plots that were generated and realised that there
was no command-line option to modify the configuration value loaded from the YAML (`topostats/default_config.yaml` or
user specified with `--config-file/-c`).

This means that currently if people want the `all` image set then they have to use `topostats create-config` to create a
configuration file, modify the value of `image_set` and run the analysis with `topostats -c new_config.yaml`.

It was very quick and easy to add this option so I could check the output and as its been done I felt it should be
incorporated into TopoStats.