refactor: image_flattened > image
In working through adding the `topostats grains` entry point I was confused why `.topostats` files had `image` which
contained the flattened image but all the processing stages after `Filters` used `image_flattened`. After checking with
@SylviaWhittle we have opted to make things consistent across the file output and the processing.

Updates tests in light of these changes, previously the result of `loadscans.get_data()` left `img_dict` as a dictionary
of the _data_ but to align with other scan types we actually want a nested dictionary with the keys as the filenames
then the data (whether that is a single scan from most raw data or the dictionary that `.topostats` files hold).

This raises (again as I've asked a similar question before before) the disconnect between a `topostats` object internal
to TopoStats and as stored in the HDF5 file format and the value returned by AFMReader. As can be seen
[here](https://github.com/AFM-SPM/AFMReader/blob/022dcf286914c23a30da42e4ea401aa577b0b193/AFMReader/topostats.py#L55)
for convenience the `image`, `pixel_to_nm_scaling` are extracted from `data` and returned as part of a Tuple along with
the `data` from which it was extracted. This _might_ be convenient for users but I see no reason why we shouldn't return
just `data` and then users can access these values via `data["image"]` and `data["pixel_to_nm_scaling"]`. This would
mean the result of loading a `.topostats` object matches the interal representation and we can remove some logic and
wrangling that has been introduced in this PR to sort that out.