bug: correct dest in argparse for splining
As @SylivaWhittle noticed the `dest` arguments for the splining arguments on the command line entry-points didn't align
with the parameter names in `default_config`. This meant that if used the arguments would not be updated.

These are now aligned and the test parameters provided are different from those in the `default_config.yaml`.