feature: Introduce TopoStats class and saving to HDF5 .topostats
As we look to move towards restructuring the `.topostats` HDF5 file I've taken the logical extension to the introduction of `ImageGrainCrops` dataclass and the nested dataclass/class by introducing a `TopoStats` dataclass which defines the top-level attributes and structure that we wish to have in `.topostats` files.

One of the disconnects at the moment is that when we run the whole processing pipeline we load whatever image using AFMReader which returns `image` and `pixel_to_nm_scale`. This is ok if we're running the whole pipeline but when it comes to later stages and attempting to load `.topostats` files which can be passed into `Grains`/`GrainStats`/`DisorderedTracing`/etc. then we will be loading a `.topostats` file which is stored as HDF5, initially loaded as a dictionary but we will want to convert all previous steps of processing to `ImageGrainCrops` and its nested structure.

That means that `LoadScans` would return different things which isn't a good idea, it should consistently return the _same_ thing each time it loads files with `LoadScans.get_data()`, to which end I think it should create a dictionary of `TopoStats` objects. If data is being loaded from anything other than `.topostats` files then each value of the
dictionary, which is a `TopoStats` object will have the following attributes...

- `image_original`
- `pixel_to_nm_scale`
- `filename`
- `topostats_version` (populated on the fly from the current version)
- `img_path`

...but the `image_grain_crops` attribute will be empty.

If loading a `.topostats` file then `image_grain_crops` will hold the results of all previously processed data.

Nothing has _yet_ changed in how files are saved, but I'm (hopefully) setting things up for these changes to be made but breaking it down into a smaller commits/pull requests to ease the review process.