TopoStats classes - Grains
The `Grains` class now works with `TopoStats` classes, however...because `GrainCrops` was used in `TopoStats` and this work meant `TopoStats` was used by `Grains` we introduced a circular dependency which Python, reasonably,  complains about. The solution has been to move the class definitions to their own modules `topostats.classes`, but that wasn't without some issues since there are static methods of the `Grains` class that were used _within_ `GrainCrop`. For now these have been moved to the `utils` module and I've started writing tests for them (as they didn't appear to have any).

As a consequence this commit has a lot of things moving around which _will_ make it a pain to review, but hopefully this will be worth it.

For now the whole test suite does _not_ pass all tests because the integration tests where the pipeline is run
end-to-end fails. No attempt has been made to correct this yet because ultimately we would like to simply update the `TopoStats` objects and pass them around and that will only be addressed once each processing step/class has been refactored to work with these.

Subsequent modules should be a little easier to refactor now that the circular dependencies have been broken.