TopoStats classes - GrainStats
Switches `GrainStats` to take the `TopoStats` object as an argument and extract the `ImageGrainCrops.GrainCropDirection.crops` (be that `above` or `below`) and calculates the statistics from the returned dictionary.

Tests are updated and passed for this module alone, integration tests still fail and will be addressed after all modules are updated.