WIP TopoStats classes - Nodestats adding attrs to GrainCrop
- Move things around in resources so they are better organised.
- Disable parallel testing, its a pain when developing and running individual tests, too much setup time.

Need to add a bunch of attributes from disordered tracing to `GrainCrop` objects so that they are available....

`disordered_tracing_direction_data` which is a dictionary. Whether this should get passed into the `NodeStats` class I'm
unsure, that adds a lot of work passing things around and I would like to get the basic idea of working with GrainCrop
objects working and only then dig really deep into replacing things.

To which end can probably undo some of the changes in this PR that are made to the nodeStats class as it is
`notestats_image()` which instantiates the class with the correct things.