tests|ci: Remove configuration to always run tests in parallel
Found this to be a hinderance during development as setting up workers can take longer than running small batches of tests.

Therefore removing this option the CI already uses `--numprocesses=logical` and I think we as developers can now to run `pytest -n #` when we want to run things in parallel.