tests(processing): nodestats classes and tests
- Introduces [pytest-profiling](https://pypi.org/project/pytest-profiling/) as a test dependency so we can profile
tests. Introduced because `nodestats` was taking a looooong time to run and its because of long calls to `networkx` that
are required to get edges/angles.
- Adds `catenane_topostats` and `minicircle_small_topostats` fixtures used in `test_run_nodestats()`.
- Tests `run_nodestats`, another step in the right direction of modularising and adding entry points. Note that the
`catenane` image has 41 nodes which is one of the reason tests take so long!
- Corrects asssertions in `test_run_grains()` to be madea against `topostats_object` attributes rather than pulling out
and assigning to `imagegraincrops`.
- Rounds out the `Nodes` class with documentation and attributes.
- Switches to assessing whether disordered tracing worked by comparing the shape of the dataframe to `(0, 0)` which is
the shape of an empty dataframe. Previously this test was done against `if disordered_trace_grainstats is not None` but
as the following shows a `pd.DataFrame()` can't be used for truthiness as is normally the case in Python as an empty
dataframe is "something" so the test wasn't doing what was expected.

```
pd.DataFrame() is None
False
pd.DataFrame is not None
True
```

It is worth noting that there are some Warnings raised, these were noticed when testing for equality of Nodestats and
I've not got the time to investigate these fully, comments have been left in place so we can address in the future and
I'll make an issue for these too.