feature(io): dict_to_topostats() function
Required because loading `.topostats` objects from HDF5 AFMReader returns dictionaries. This is ok and I think for now
we should not change this as it makes AFMReader very general and of use to others, but internally when we are switching
to `TopoStats` classes for all the processing each entry point that loads a `.topostats` file requires a `TopoStats`
object so we _have_ to convert these on loading.