feature: Add *_to_dict() methods to new classes
Add `_to_dict()` methods to each of the following classes...

- `MatchedBranch`
- `Molecule`
- `Node`
- `OrderedTrace`

...and ensures these are written to HDF5.

Adds dummy objects to `tests/contest.py` and tests the methods work via `tests/test_classes.py`.

Currently the types of many of these are _wrong_ because I don't know what they actually represent, that doesn't really
matter for the testing though which uses dictionary comprehension and handles any type.

Key is that the `GrainCrop.grain_crop_to_dict()` method now works with all of the additional attributes so we can write
the full `TopoStats` object to HDF5 which is required for on-going test development of the remaining `OrderedTrace`,
`Splining` and `Curvature` so we can write intermediary `.topostats` objects which we can load for tests (instead of
running the whole processing pipeline from the start).

This is however also **vital** to the additional entry-points (aka "swiss-army knife") work so we can write `.topostats`
objects with all of the data upto a given point and load it in the future (previous commit e731084665 added the
necessary `dict_to_topostats()` function for converting the HDF5-based dictionaries to `TopoStats` objects).