feature(classes): Pydantic classes and pickling
Moves to [Pydantic  dataclasses](https://docs.pydantic.dev/latest/concepts/dataclasses/) for stricter data validation.

This means we can pickle `TopoStats` objects which is useful because in the test suite we don't want to run the whole
pipeline when we want to test e.g. `Nodestats`. As a consequence we now have pickles which are loaded as
`pytest.fixtures` (from `tests/conftest.py`) rather than lines of code within tests themselves that save and modify
`.npy`/`.pkl` files.

There are therefore three sets of pickles...

- `minicircle_small`
- `catenanes`
- `rep_int`

...at different stages...

- `_post_grainstats`
- `_post_disordered_tracing`
- `_post_nodestats`

...and we will develop additional fixtures for...

- `_post_ordered_tracing`
- `_post_curvature`
- `_post_splining` (optional, not required at the moment as no subsequent processing is done after this)

A slight disconnect might arise from how these pickles were created, at the moment it is code in a `.py` file on @ns-rse
computer. @ns-rse will look at adding this as an additional script in the repository, but as more work is required its
not included at the moment.

This now allows me to finish of re-factoring and writing the integration test for `ordered-tracing`.