refactor14: tidying return values and functions
Starts removing the returning of objects and dataframes by various functions and removing the passing of these
around. Not yet done for `topostats/processing.py` though and not complete.

In particular a lot of work is required to get the plotting back into shape (as well as addressing a hang in the
splining tests which I'm yet to get to the bottom of).