fix(config): remove `default_` from output config file
A minor bug was found which arises when no `filename` is provided as an argument to `topostats create_config`.

When this is the case the `filename` is set conditional on the type of config which meant that the `default`
configuration filename `default_config.yaml` was used.

The documentation and help indicate that if no argument is given it will be written to `config.yaml` and so two tests
were failing.

This is corrected and tests have been expanded to cover the additional configuration writing options for `var_to_label`
and `mplstyle`.